+++ MyApps/Entities/Payement.cs
@@ -8,7 +8,7 @@ namespace Entities
     {
         public int Id { get; set; }
         public string Ref { get; set; }
-        public Decimal Price { get; set; }
+        public double Price { get; set; }
         public DateTime Payement_date { get; set; }
       
     }
